different contest mode corresponding to different export format
